split the network policies code to its own file
